MQ/Python/requirements.txt
python-dotenv==0.20.0
azure-eventhub==5.11.2
paho-mqtt==1.6.1